+++ src/components/Sidebar/SortableBlockItem.tsx
@@ -3,9 +3,16 @@ import { CSS } from "@dnd-kit/utilities";
 import { FiMoreVertical } from "react-icons/fi";
 import type { ReactNode } from "react";
 
+interface Block {
+  id: string;
+  type: string;
+  content: string;
+  level?: number;
+}
+
 interface SortableBlockItemProps {
   id: string;
-  block: any;
+  block: Block;
   icon: ReactNode;
   title: string;
   onClick: () => void;
