+++ src/types/editor-context.types.ts
@@ -1,0 +1,6 @@
+import { Editor } from '@tiptap/react';
+
+export interface EditorContextType {
+  editor: Editor | null;
+  setEditor: (editor: Editor | null) => void;
+}
